Database initialization should be ON by default
If initdb=0, then the database is not created or initialized, which
is very confusing on first run (install).  If initdb=1, then first
run is good, but the re-runs will fail due to the database already
being present.

For now, setting initdb=1 to avoid confusion for the first time
users.  Later (soon!) will add a check for the database presence and
some sane behavior.